+++ src/pages/dataset/uiSchema.ts
@@ -21,6 +21,7 @@ const datasetUiSchema = {
     "name",
     "data_accessibility",
     "data_access_link",
+    "data_access_date",
     "experiment_id",
     "description",
     "temporal_coverage",
@@ -50,6 +51,11 @@ const datasetUiSchema = {
   data_access_link: {
     ...halfWidthStyle,
     "ui:placeholder": "https://doi.org/10.25921/...",
+  },
+  data_access_date: {
+    ...halfWidthStyle,
+    "ui:placeholder": "YYYY-MM-DD",
+    "ui:descriptionModal": true,
   },
   description: textAreaWidget,
   // Project ID - hidden, auto-inferred from parent project